RIH MAL
This is a personal project based on a friend's idea. It has been some time I started working on this but the situation in Myanmar gets complicated and I didn't finish it. Just a few hours ago, a similar game was also hosted by one of the developers in Myanmar. That encouraged me to finish up no matter what. Enjoy!